gptprio: ensure disk is available before pre-mount phase
The 'wait_for_dev' functions, despite their misleading name, do not wait
or otherwise ensure any ordering between their usage and when the
devices become available. It turns out that they simply disable
systemd's default timeout for devices and adds a dependency to
initrd.target. The two combined help ensure that the initrd may
mysteriously hang forever given an error but fail with race conditions
when things are otherwise working.

The mixture of dracut and systemd continues to be one of the worst
designed init systems available: instead of one init system you get
three, two of which are written in sh-compatible shells scripts. It
would be better to convert the little bit of dracut specific modules in
bootengine to use systemd exclusively but don't have the time. Also
eventually gptprio will be moved from the initrd to a bootloader anyway.

For now replace wait_for_dev with a systemd generator that requires the
devices to be available before the pre-mount dracut phase runs, where
the main gptprio code hooks in. Without this it was possible for
pre-mount to run too early, failing, and falling back via a reboot.

The new generator doesn't bother with root=gptprio: since those images
are long gone. Hard fail if that arg is encountered.